Move StratumFromTraits.
