+++ lib/ModelInterfaces/WithStrata.ts
@@ -1,4 +1,4 @@
-import StratumFromTraits from "./StratumFromTraits";
+import StratumFromTraits from "../Models/StratumFromTraits";
 
 export default interface WithStrata<Traits> {
     getOrCreateStratum(id: string): StratumFromTraits<Traits>;
