+++ lib/Models/StratumFromTraits.ts
